form producto

+++ herreria/models.py
@@ -5,8 +5,8 @@ class producto(models.Model):
     nombre = models.CharField(max_length=50, unique=True, blank=False,)
     imagen = models.ImageField(upload_to='productos/', null=True, blank=True)
     descripcion = models.TextField()
-    precio = models.FloatField()
-    stock = models.IntegerField()
+    precio = models.DecimalField(max_digits=10, decimal_places=2)
+    stock = models.PositiveBigIntegerField()
 
     def __str__(self):
         return self.nombre
